Unify Serve func
